+++ src/App.tsx
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/lines-between-class-members */
 import React from 'react';
 import './App.scss';
+import { AppState } from './types/AppState';
+import { Clock } from './components/Clock';
 
 function getRandomName(): string {
   const value = Date.now().toString().slice(-4);
@@ -10,41 +12,6 @@ function getRandomName(): string {
 
 function getCurrentTime() {
   return new Date().toUTCString().slice(-12, -4);
-}
-
-interface ClockProps {
-  clockName: string;
-  currentTime: string;
-}
-
-export class Clock extends React.Component<ClockProps> {
-  componentDidUpdate(prevProps: ClockProps) {
-    if (prevProps.clockName !== this.props.clockName) {
-      // eslint-disable-next-line no-console
-      console.debug(
-        `Renamed from ${prevProps.clockName} to ${this.props.clockName}`,
-      );
-    }
-  }
-
-  render() {
-    const { clockName, currentTime } = this.props;
-
-    return (
-      <div className="Clock">
-        <strong className="Clock__name">{clockName}</strong>
-        {' time is '}
-        <span className="Clock__time">{currentTime}</span>
-      </div>
-    );
-  }
-}
-
-interface AppState {
-  clockName: string;
-  currentTime: string;
-  hasClock: boolean;
-  isConsoleLogRunning: boolean;
 }
 
 export class App extends React.Component<{}, AppState> {
