+++ src/components/Clock.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { ClockProps } from '../types/Clock';
+
+export class Clock extends React.Component<ClockProps> {
+  componentDidUpdate(prevProps: ClockProps) {
+    if (prevProps.clockName !== this.props.clockName) {
+      // eslint-disable-next-line no-console
+      console.debug(
+        `Renamed from ${prevProps.clockName} to ${this.props.clockName}`,
+      );
+    }
+  }
+
+  render() {
+    const { clockName, currentTime } = this.props;
+
+    return (
+      <div className="Clock">
+        <strong className="Clock__name">{clockName}</strong>
+        {' time is '}
+        <span className="Clock__time">{currentTime}</span>
+      </div>
+    );
+  }
+}
