+++ src/types/AppState.ts
@@ -1,0 +1,6 @@
+export interface AppState {
+  clockName: string;
+  currentTime: string;
+  hasClock: boolean;
+  isConsoleLogRunning: boolean;
+}
