+++ src/types/Clock.ts
@@ -1,0 +1,4 @@
+export interface ClockProps {
+  clockName: string;
+  currentTime: string;
+}
